编辑、删除用户页面制作与对接完成

+++ gym-base-parent/gym-base-web/src/main/java/com/gym/web/sys_user/service/SysUserService.java
@@ -14,7 +14,7 @@ public interface SysUserService extends IService<SysUser> {
     // 新增
     void addUser(SysUser sysUser);
     // 删除
-    void deleteUser(long userId);
+    void deleteUser(Integer userId);
     // 修改
     void editUser(SysUser sysUser);
     // 列表
